signal changes

+++ src/olympia/promoted/signals.py
@@ -1,6 +1,8 @@
 from django.db import models
 from django.db.models.signals import ModelSignal
 
+from olympia import amo
+from olympia.constants.applications import APP_USAGE
 from olympia.constants.promoted import PROMOTED_GROUP_CHOICES
 
 from .models import (
@@ -30,31 +32,42 @@ def promoted_addon_to_promoted_addon_promotion(
     elif instance.pk:
         PromotedAddonPromotion.objects.filter(addon=instance.addon).delete()
 
-    # If the addon has previously been approved
-    # (see test_unbadged_non_prereview_promoted_can_be_disabled_and_deleted):
-    # (i.e has existing approvals) for the current group:
-    # If promotedaddon's application_id is not None,
-    # delete the PromotedAddonVersions that are not the current application_id
-    # and make sure the current application_id exists.
-    # Otherwise, it should be available for all applications.
-    # This should mirror the behaviour of PromotedAddon's all_applications()
-    # when used by approved_applications.
+    channel = instance.addon.current_version.channel
+    prereview = (
+            (channel == amo.CHANNEL_LISTED and promoted_group.listed_pre_review)
+            or (channel == amo.CHANNEL_UNLISTED and promoted_group.unlisted_pre_review)
+        )
 
     existing_approval = PromotedApproval.objects.filter(
         version=instance.addon.current_version, group_id=promoted_group.group_id
     )
 
-    if existing_approval.exists():
-        if instance.application_id:
-            print(instance.group_id, instance.application_id)
-            PromotedAddonVersion.objects.filter(
-                version=instance.addon.current_version
-            ).exclude(application_id=instance.application_id).delete()
+    # 1. If the group is not prereviewed, approve for all apps.
+    if not prereview:
+        for app in APP_USAGE:
             PromotedAddonVersion.objects.update_or_create(
                 version=instance.addon.current_version,
                 promoted_group=promoted_group,
-                application_id=instance.application_id,
+                application_id=app.id,
             )
+    # 2. If the addon has previously been approved
+    # (i.e has existing approvals) for the current group
+    # and promotedaddon's application_id is not None:
+    # a) delete the PromotedAddonVersions that are not the
+    # current application_id, and
+    # b) make sure the current application_id exists.
+    # Otherwise, it should be available for all applications.
+    # This should mirror the behaviour of PromotedAddon's 
+    # all_applications() when used by approved_applications.
+    elif existing_approval.exists() and instance.application_id:
+        PromotedAddonVersion.objects.filter(
+            version=instance.addon.current_version
+        ).exclude(application_id=instance.application_id).delete()
+        PromotedAddonVersion.objects.update_or_create(
+            version=instance.addon.current_version,
+            promoted_group=promoted_group,
+            application_id=instance.application_id,
+        )
 
 
 def promoted_approval_to_promoted_addon_version(
